+++ CHANGES.rst
@@ -7,8 +7,9 @@ LATEST
 - Added `max_length` parameter to `Text`.
 - Added support for page up/down in `TextBox`.
 - Added optional scroll bars to `MultiColumnListBox`.
-- Added `file_filter` parameter to `FileBrowser`
-- Added `wait_for-input` method to `Screen`
+- Added `file_filter` parameter to `FileBrowser`.
+- Added `wait_for-input` method to `Screen`.
+- Added optional `theme` parameter to `PopupDialog`.
 - Improved performance of double-buffering.
 
   - NOTE: Drawing off-screen with a large scrolling buffer is no longer supported (as it wasn't